add agent base;add ConfigParserBase
